avoid assert_batches duplication
